feat: drop support prefixed env vars + default service account file
